Fixed Coord.dist.

+++ src/haven/Coord.java
@@ -144,8 +144,8 @@ public class Coord implements Comparable<Coord>, java.io.Serializable {
     }
 	
     public double dist(Coord o) {
-	int dx = o.x - x;
-	int dy = o.y - y;
+	long dx = o.x - x;
+	long dy = o.y - y;
 	return(Math.sqrt((dx * dx) + (dy * dy)));
     }
 }
